refactor(eval-workflows): evaluation-pipeline 610 行按 5 个 cluster 拆为 orchestrator + 4 helper
把 evaluation-pipeline.ts 按职责拆成主文件(orchestrator + options)+ 同级 evaluation-pipeline/ 目录下 4 个 helper:
- run-state.ts: EvaluationRunState 生命周期(init / finalizeSuccess / persistSuccess / persistFailed)
- test-set-hash.ts: 测试集水印 hash(spec §7.1)
- preflight-warnings.ts: power + isolation 结构性预警
- report-finalize.ts: report 后置分析装配(analysis / coverage / gapReports / blind)

主文件从 610 行收敛到约 280 行,纯粹是「执行+收尾」时序编排。helper 之间无环依赖,每块都是单向流入 orchestrator。

兼容 re-export 保留 buildPowerWarnings / buildIsolationWarnings / _computeTestSetHashForTest,避免 test 与 run-evaluation.ts:275 动态 import 路径改动。无 BREAKING-COMPARABILITY:测量学红区文件零改动,Report schema / judge prompt / observe LLM 复盘 prompt / Bootstrap / α / length-debias toggle 全不动,所有改动是搬代码不改语义。
